* Simplify ArchaiusModule to make it less dependent on being customized using DI * Remove the concept of library overrides provided as guice bindings (it's overly complex and not necessary) * Get rid of the ArchaiusConfiguration class * Get rid of the ConfigSeeder concept as it isn't necessary * Fix ConfigProxyFactory's use of the deprecated ProxyFactory
