--- archaius2-api/src/main/java/com/netflix/archaius/api/inject/ApplicationLayer.java
@@ -1,15 +1,0 @@
-package com.netflix.archaius.api.inject;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-import javax.inject.Qualifier;
-
-@Qualifier
-@Target({ElementType.TYPE, ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface ApplicationLayer {
-
-}
