+++ archaius2-archaius1-bridge/src/test/java/com/netflix/archaius/bridge/AbstractConfigurationBridgeTest.java
@@ -20,9 +20,8 @@ import com.google.inject.Key;
 import com.netflix.archaius.api.Config;
 import com.netflix.archaius.api.annotations.ConfigurationSource;
 import com.netflix.archaius.api.config.SettableConfig;
-import com.netflix.archaius.guice.ArchaiusModule;
-import com.netflix.archaius.guice.ConfigSeeders;
 import com.netflix.archaius.api.inject.RuntimeLayer;
+import com.netflix.archaius.guice.ArchaiusModule;
 import com.netflix.archaius.visitor.PrintStreamVisitor;
 import com.netflix.config.AggregatedConfiguration;
 import com.netflix.config.ConfigurationManager;
@@ -57,7 +56,6 @@ public class AbstractConfigurationBridgeTest {
             install(new ArchaiusModule());
             
             bind(SomeClient.class).asEagerSingleton();
-            ConfigSeeders.bind(binder(), properties, RuntimeLayer.class);
         }
     }
