+++ archaius2-guice/src/main/java/com/netflix/archaius/guice/ArchaiusModule.java
@@ -15,299 +15,261 @@
  */
 package com.netflix.archaius.guice;
 
+import java.util.Properties;
 import java.util.Set;
-import java.util.concurrent.atomic.AtomicInteger;
 
 import com.google.inject.AbstractModule;
+import com.google.inject.Binding;
+import com.google.inject.Injector;
+import com.google.inject.Key;
 import com.google.inject.Provides;
+import com.google.inject.ProvisionException;
 import com.google.inject.Scopes;
 import com.google.inject.Singleton;
 import com.google.inject.matcher.Matchers;
 import com.google.inject.multibindings.Multibinder;
-import com.netflix.archaius.api.Config;
-import com.netflix.archaius.api.ConfigListener;
-import com.netflix.archaius.api.ConfigLoader;
 import com.netflix.archaius.ConfigProxyFactory;
-import com.netflix.archaius.api.ConfigReader;
 import com.netflix.archaius.DefaultConfigLoader;
+import com.netflix.archaius.DefaultDecoder;
 import com.netflix.archaius.DefaultPropertyFactory;
+import com.netflix.archaius.api.CascadeStrategy;
+import com.netflix.archaius.api.Config;
+import com.netflix.archaius.api.ConfigLoader;
+import com.netflix.archaius.api.ConfigReader;
+import com.netflix.archaius.api.Decoder;
 import com.netflix.archaius.api.PropertyFactory;
-import com.netflix.archaius.cascade.NoCascadeStrategy;
 import com.netflix.archaius.api.config.CompositeConfig;
-import com.netflix.archaius.config.DefaultCompositeConfig;
-import com.netflix.archaius.config.DefaultCompositeConfig.Builder;
-import com.netflix.archaius.config.DefaultSettableConfig;
-import com.netflix.archaius.config.EnvironmentConfig;
 import com.netflix.archaius.api.config.SettableConfig;
-import com.netflix.archaius.config.SystemConfig;
 import com.netflix.archaius.api.exceptions.ConfigException;
-import com.netflix.archaius.api.inject.ApplicationLayer;
-import com.netflix.archaius.api.inject.DefaultsLayer;
 import com.netflix.archaius.api.inject.LibrariesLayer;
 import com.netflix.archaius.api.inject.RemoteLayer;
 import com.netflix.archaius.api.inject.RuntimeLayer;
+import com.netflix.archaius.cascade.NoCascadeStrategy;
+import com.netflix.archaius.config.DefaultCompositeConfig;
+import com.netflix.archaius.config.DefaultSettableConfig;
+import com.netflix.archaius.config.EnvironmentConfig;
+import com.netflix.archaius.config.MapConfig;
+import com.netflix.archaius.config.SystemConfig;
 import com.netflix.archaius.interpolate.ConfigStrLookup;
 import com.netflix.archaius.readers.PropertiesConfigReader;
 
 /**
- * Guice module with default bindings to enable Config injection and 
- * configuration mapping/binding in a Guice based application.
+ * Guice Module for setting up archaius and making it's components injectable and provides the following
+ * functionality
  * 
- * By default this module will create a top level Config that is a 
- * CompositeConfig of the following layers,
- * RUNTIME     - properties set from code
- * OVERRIDE    - properties loaded from a remote source
- * SYSTEM      - System properties
- * ENVIRONMENT - Environment properties
- * APPLICATION - Configuration loaded by the application
- * LIBRARIES   - Configuration loaded by libraries used by the application
- * DEFAULTS    - Defaults that can be set from code
+ * <ul>
+ * <li>Configuration Proxy</li>
+ * <li>Injectable Config</li>
+ * <li>Configuration mapping</li>
+ * </uL>
  * 
- * Runtime properties may be set by either injecting and calling one of the
+ * Note that this module has state and should therefore only be installed once.  It should only
+ * be installed by applications and not by libraries depending on archaius that.  Such libraries 
+ * should only specify a bindger.requriesBinding(Config.class).  
+ * 
+ * Archaius runs with the following override structure
+ *  RUNTIME     - properties set from code
+ *  REMOTE      - properties loaded from a remote source
+ *  SYSTEM      - System properties
+ *  ENVIRONMENT - Environment properties
+ *  APPLICATION - Configuration loaded by the application
+ *  LIBRARIES   - Configuration loaded by libraries used by the application
+ * 
+ * Runtime properties may be set in code by injecting and calling one of the
  * setters for,
- *  @RuntimeLayer SettableConfig config
- * or at startup by specifying a multibinding
- *  Multibinder.newSetBinder(binder(), ConfigResolver.class, RuntimeLayer.class)...
+ *  {@literal @}RuntimeLayer SettableConfig config
  *  
- * Override properties may be set by either injecting 
- *  @OverrideLayer CompositeConfig config
- * or at startup by specifying a multibinding
- *  Multibinder.newSetBinder(binder(), ConfigResolver.class, OverrideLayer.class)...
+ * A remote configuration may be specified by binding to {@literal @}RemoteLayer Config
+ * When setting up a remote configuration that need access to archaius's Config
+ * make sure to inject the qualifier {@literal @}Raw otherwise the injector will fail
+ * with a circular dependency error.  Note that the injected config will have 
+ * system, environment and application properties loaded into it.
  * 
- * Application properties will be loaded automatically from resources with the base 
- * name indicated by the binding to @ApplicationLayer String.  The default value
- * is 'application' be can be changed via a binding override.
- * 
- * @author elandau
+ * <code>
+ * public class FooRemoteModule extends AbstractModule {
+ *     {@literal @}Override
+ *     protected void configure() {}
+ *     
+ *     {@literal @}Provides
+ *     {@literal @}RemoteLayer
+ *     Config getRemoteConfig({@literal @}Raw Config config) {
+ *         return new FooRemoteConfigImplementaiton(config);
+ *     }
+ * }
+ * </code>
  */
 public final class ArchaiusModule extends AbstractModule {
+    private static final String DEFAULT_CONFIG_NAME     = "application";
     
-    private static final String RUNTIME_LAYER_NAME              = "RUNTIME";
-    private static final String REMOTE_LAYER_NAME               = "REMOTE";
-    private static final String SYSTEM_LAYER_NAME               = "SYSTEM";
-    private static final String ENVIRONMENT_LAYER_NAME          = "ENVIRONMENT";
-    private static final String APPLICATION_OVERRIDE_LAYER_NAME = "APPLICATION_OVERRIDE";
-    private static final String APPLICATION_LAYER_NAME          = "APPLICATION";
-    private static final String LIBRARIES_LAYER_NAME            = "LIBRARIES";
-    private static final String DEFAULTS_LAYER_NAME             = "DEFAULTS";
+    private static final String RUNTIME_LAYER_NAME      = "RUNTIME";
+    private static final String REMOTE_LAYER_NAME       = "REMOTE";
+    private static final String SYSTEM_LAYER_NAME       = "SYSTEM";
+    private static final String ENVIRONMENT_LAYER_NAME  = "ENVIRONMENT";
+    private static final String APPLICATION_LAYER_NAME  = "APPLICATION";
+    private static final String LIBRARIES_LAYER_NAME    = "LIBRARIES";
     
-    private static final AtomicInteger idCounter = new AtomicInteger();
+    static {
+        System.setProperty("archaius.default.configuration.class",      "com.netflix.archaius.bridge.StaticAbstractConfiguration");
+        System.setProperty("archaius.default.deploymentContext.class",  "com.netflix.archaius.bridge.StaticDeploymentContext");
+    }
+    
+    private String                  configName           = DEFAULT_CONFIG_NAME;
+    private Config                  applicationOverrides = null;
+    private Class<? extends CascadeStrategy> cascadeStrategy = NoCascadeStrategy.class;
+
+    private final SettableConfig  runtimeLayer;
+    private final CompositeConfig remoteLayer;
+    private final CompositeConfig applicationLayer;
+    private final CompositeConfig librariesLayer;
+    private final CompositeConfig rawConfig;
+
+    public ArchaiusModule() {
+        this.runtimeLayer     = new DefaultSettableConfig();
+        this.applicationLayer = new DefaultCompositeConfig();
+        this.librariesLayer   = new DefaultCompositeConfig();
+        this.remoteLayer      = new DefaultCompositeConfig();
+        try {
+            this.rawConfig       = DefaultCompositeConfig.builder()
+                    .withConfig(RUNTIME_LAYER_NAME,      runtimeLayer)
+                    .withConfig(REMOTE_LAYER_NAME,       remoteLayer)
+                    .withConfig(SYSTEM_LAYER_NAME,       SystemConfig.INSTANCE)
+                    .withConfig(ENVIRONMENT_LAYER_NAME,  EnvironmentConfig.INSTANCE)
+                    .withConfig(APPLICATION_LAYER_NAME,  applicationLayer)
+                    .withConfig(LIBRARIES_LAYER_NAME,    librariesLayer)
+                    .build();
+        } catch (ConfigException e) {
+            throw new ProvisionException("Error creating raw configuration", e);
+        }
+    }
+    
+    public ArchaiusModule withConfigName(String value) {
+        this.configName = value;
+        return this;
+    }
+    
+    public ArchaiusModule withApplicationOverrides(Properties prop) throws ConfigException {
+        return withApplicationOverrides(MapConfig.from(prop));
+    }
+    
+    public ArchaiusModule withApplicationOverrides(Config config) throws ConfigException {
+        this.applicationOverrides = config;
+        return this;
+    }
+    
+    public ArchaiusModule withCascadeStrategy(Class<? extends CascadeStrategy> cascadeStrategy) {
+        this.cascadeStrategy = cascadeStrategy;
+        return this;
+    }
     
     @Override
     protected void configure() {
         bindListener(Matchers.any(), new ConfigurationInjectingListener());
-
+        
+        bind(CascadeStrategy.class).to(cascadeStrategy);
+        
         Multibinder.newSetBinder(binder(), ConfigReader.class)
             .addBinding().to(PropertiesConfigReader.class).in(Scopes.SINGLETON);
-        
-        bind(ArchaiusConfiguration.class).to(OptionalArchaiusConfiguration.class);
     }
-    
+
     @Provides
     @Singleton
     @RuntimeLayer
     SettableConfig getSettableConfig() {
-        return new DefaultSettableConfig();
+        return runtimeLayer;
     }
     
-    @Provides
-    @Singleton
-    @DefaultsLayer
-    SettableConfig getDefaultsConfig() {
-        return new DefaultSettableConfig();
-    }
-    
-    @Provides
-    @Singleton
-    @ApplicationLayer 
-    CompositeConfig getApplicationLayer() {
-        return new DefaultCompositeConfig();
-    }
-
     @Provides
     @Singleton
     @LibrariesLayer
     CompositeConfig getLibrariesLayer() {
-        return new DefaultCompositeConfig();
+        return librariesLayer;
     }
     
     @Provides
     @Singleton
-    @RemoteLayer
-    CompositeConfig getRemoteLayer() {
-        return new DefaultCompositeConfig();
+    @Raw
+    Config getRawConfig() {
+        return rawConfig;
     }
     
     @Provides
     @Singleton
-    @RemoteLayer
-    Config getRemoteLayer(@RemoteLayer CompositeConfig config) {
-        return config;
+    Config getConfig(ConfigLoader loader, Injector injector) throws Exception {
+        if (applicationOverrides != null) {
+            applicationLayer.addConfig("overrides", applicationOverrides);
+        }
+        
+        this.applicationLayer.addConfig("loaded",  loader
+            .newLoader()
+            .load(configName));
+
+        // load any runtime overrides
+        Binding<Config> binding = injector.getExistingBinding(Key.get(Config.class, RemoteLayer.class));
+        if (binding != null) {
+            // TODO: Ideally this should replace the remoteLayer in config but there is a bug in archaius
+            //       where the replaced layer moves to the end of the hierarchy
+            remoteLayer.addConfig("remote", binding.getProvider().get());
+        }
+        
+        return rawConfig;
     }
-    
-    /**
-     * This is the main config loader for the application.
-     * 
-     * @param config
-     * @param archaiusConfiguration
-     * @param readers
-     * @return
-     */
+        
     @Provides
     @Singleton
     ConfigLoader getLoader(
-            @RootLayer            Config config,
-            ArchaiusConfiguration archaiusConfiguration,
+            @LibrariesLayer       CompositeConfig libraries,
+            CascadeStrategy       cascadingStrategy,
             Set<ConfigReader>     readers
-            ) {
+            ) throws ConfigException {
+        
         return DefaultConfigLoader.builder()
-                .withConfigReader(readers)
-                .withDefaultCascadingStrategy(archaiusConfiguration.getCascadeStrategy())
-                .withFailOnFirst(false)
-                .withStrLookup(ConfigStrLookup.from(config))
-                .build();
-    }
-
-    /**
-     * The internal RootLayer is meant only to construct the override hierarchy and 
-     * not to load or initialize any of the layers.  Initialization of layers is done
-     * is getConfig()
-     * 
-     * @param archaiusConfiguration
-     * @param settableLayer
-     * @param overrideLayer
-     * @param applicationLayer
-     * @param librariesLayer
-     * @param defaultsLayer
-     * @return
-     * @throws ConfigException
-     */
-    @Provides
-    @Singleton
-    @RootLayer
-    Config getInternalConfig(
-            ArchaiusConfiguration archaiusConfiguration,
-            @RuntimeLayer             SettableConfig    settableLayer,
-            @RemoteLayer              Config            overrideLayer,
-            @ApplicationLayer         CompositeConfig   applicationLayer, 
-            @LibrariesLayer           CompositeConfig   librariesLayer,
-            @DefaultsLayer            SettableConfig    defaultsLayer) throws ConfigException {
-        Builder builder = DefaultCompositeConfig.builder()
-                .withConfig(RUNTIME_LAYER_NAME,              settableLayer)
-                .withConfig(REMOTE_LAYER_NAME,               overrideLayer)
-                .withConfig(SYSTEM_LAYER_NAME,               SystemConfig.INSTANCE)
-                .withConfig(ENVIRONMENT_LAYER_NAME,          EnvironmentConfig.INSTANCE);
-        
-        if (archaiusConfiguration.getApplicationOverride() != null) {
-            builder.withConfig(APPLICATION_OVERRIDE_LAYER_NAME, archaiusConfiguration.getApplicationOverride());
-        }
-        builder.withConfig(APPLICATION_LAYER_NAME,          applicationLayer)
-                .withConfig(LIBRARIES_LAYER_NAME,            librariesLayer)
-                .withConfig(DEFAULTS_LAYER_NAME,             defaultsLayer)
-                ;
-        
-        return builder.build();
+            .withConfigReader(readers)
+            .withDefaultCascadingStrategy(cascadingStrategy)
+            .withStrLookup(ConfigStrLookup.from(rawConfig))
+            .build();
     }
     
-    /**
-     * All code will ultimately inject Config to gain access to the entire 
-     * configuration hierarchy.  The empty hierarchy is created by @RootLayer
-     * but here we do the actual Configuration initialization which includes,
-     * 1.  Loading application properties
-     * 2.  Loading runtime overrides
-     * 3.  Loading override layer overrides 
-     * 
-     * @param archaiusConfiguration
-     * @param config
-     * @param applicationLayer
-     * @param remoteLayer
-     * @param runtimeLayer
-     * @param defaultsLayer
-     * @param loader
-     * @return
-     * @throws Exception
-     */
     @Provides
     @Singleton
-    public Config getConfig(
-            ArchaiusConfiguration archaiusConfiguration,
-            @RootLayer        Config            config,
-            @ApplicationLayer CompositeConfig   applicationLayer,
-            @RemoteLayer      CompositeConfig   remoteLayer,
-            @RuntimeLayer     SettableConfig    runtimeLayer,
-            @DefaultsLayer    SettableConfig    defaultsLayer,
-                              ConfigLoader      loader
-            ) throws Exception {
-        
-        // First load the single application configuration 
-        CompositeConfig compositeConfig = loader
-                .newLoader()
-                .withCascadeStrategy(NoCascadeStrategy.INSTANCE)
-                .load(archaiusConfiguration.getConfigName());
-        if (compositeConfig != null) {
-            applicationLayer.replaceConfig(APPLICATION_LAYER_NAME, compositeConfig);
-        }
-
-        // Load any defaults from code.  These defaults may be initialized as part of 
-        // conditional module loading provided by Governator.  
-        for (ConfigSeeder provider : archaiusConfiguration.getDefaultsLayerSeeders()) {
-            defaultsLayer.setProperties(provider.get(config));
-        }
- 
-        // load any runtime overrides
-        for (ConfigSeeder provider : archaiusConfiguration.getRuntimeLayerSeeders()) {
-            runtimeLayer.setProperties(provider.get(config));
-        }
- 
-        // Finally, load any cascaded configuration files for the application
-        compositeConfig = loader.newLoader().withCascadeStrategy(archaiusConfiguration.getCascadeStrategy()).load(archaiusConfiguration.getConfigName());
-        if (compositeConfig != null) {
-            applicationLayer.replaceConfig(APPLICATION_LAYER_NAME, compositeConfig);
-        }
-        
-        // Remote layers most likely need some configuration so we load them after application
-        // configuration has been loaded. 
-        for (ConfigSeeder provider : archaiusConfiguration.getRemoteLayerSeeders()) {
-            remoteLayer.addConfig("remote" + idCounter.incrementAndGet(), provider.get(config));
-        }
-        
-        for (ConfigListener listener : archaiusConfiguration.getConfigListeners()) {
-            config.addListener(listener);
-        }
-        
-        return config;
-    }
-    
-    /**
-     * Top level 'fast' property factory which is bound to the root configuration 
-     * @param root
-     * @return
-     */
-    @Provides
-    @Singleton
-    PropertyFactory getPropertyFactory(Config root) {
-        return DefaultPropertyFactory.from(root);
+    public Decoder getDecoder() {
+        return DefaultDecoder.INSTANCE;
     }
 
     @Provides
     @Singleton
-    ConfigProxyFactory getProxyFactory(ArchaiusConfiguration config, PropertyFactory factory) {
-        return new ConfigProxyFactory(config.getDecoder(), factory);
-    }
-    
-    @Override
-    public String toString() {
-        return getClass().getSimpleName();
-    }
-    
-    @Override
-    public boolean equals(Object obj) {
-        return ArchaiusModule.class.equals(obj.getClass());
+    PropertyFactory getPropertyFactory(Config config) {
+        return DefaultPropertyFactory.from(config);
     }
 
+    @Provides
+    @Singleton
+    ConfigProxyFactory getProxyFactory(Decoder decoder, PropertyFactory factory) {
+        return new ConfigProxyFactory(decoder, factory);
+    }
+    
     @Override
     public int hashCode() {
-        return ArchaiusModule.class.hashCode();
+        return getClass().hashCode();
+    }
+
+    /**
+     * equals() on Module's is used by Guice to dedup bindings in the event that the 
+     * same module is installed more than once as modules are assumed to not be 
+     * stateful.  ArchaiusModule however is stateful in that much of its configuration
+     * is supplied externally to the Guice injector.  Also, initializing the configuration
+     * subsystem is an application concern and as such ArchaiusModule would only 
+     * be installed once by the application itself.
+     */
+    @Override
+    public boolean equals(Object obj) {
+        if (this == obj)
+            return true;
+        if (obj == null)
+            return false;
+        if (getClass() != obj.getClass())
+            return false;
+
+        throw new ProvisionException("Only one ArchaiusModule may be installed");
     }
 
 }
