+++ archaius2-guice/src/main/java/com/netflix/archaius/guice/Raw.java
@@ -1,4 +1,4 @@
-package com.netflix.archaius.api.inject;
+package com.netflix.archaius.guice;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,9 +7,13 @@ import java.lang.annotation.Target;
 
 import javax.inject.Qualifier;
 
-@Qualifier
+/**
+ * To be used only for components meant to extend Archaius's functionality via bindings
+ * where injecting a named Config will result in a circular dependency.  
+ */
 @Target({ElementType.TYPE, ElementType.FIELD, ElementType.PARAMETER, ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface DefaultsLayer {
+@Qualifier
+@interface Raw {
 
 }
