+++ archaius2-guice/src/test/java/com/netflix/archaius/guice/ProxyTest.java
@@ -10,13 +10,13 @@ import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.Key;
 import com.google.inject.Provides;
-import com.google.inject.util.Modules;
 import com.netflix.archaius.ConfigProxyFactory;
 import com.netflix.archaius.api.annotations.Configuration;
 import com.netflix.archaius.api.annotations.DefaultValue;
-import com.netflix.archaius.config.MapConfig;
 import com.netflix.archaius.api.config.SettableConfig;
+import com.netflix.archaius.api.exceptions.ConfigException;
 import com.netflix.archaius.api.inject.RuntimeLayer;
+import com.netflix.archaius.config.MapConfig;
 
 public class ProxyTest {
     public static interface MyConfig {
@@ -42,19 +42,17 @@ public class ProxyTest {
     }
     
     @Test
-    public void testConfigWithNoPrefix() {
-        Injector injector = Guice.createInjector(Modules
-            .override(new ArchaiusModule())
-            .with(new AbstractModule() {
+    public void testConfigWithNoPrefix() throws ConfigException {
+        Injector injector = Guice.createInjector(
+            new ArchaiusModule()
+                    .withApplicationOverrides(MapConfig.builder()
+                            .put("integer", 1)
+                            .put("string", "bar")
+                            .put("subConfig.integer", 2)
+                            .build()),
+            new AbstractModule() {
                 @Override
                 protected void configure() {
-                    ConfigSeeders.bind(binder(), 
-                            MapConfig.builder()
-                                .put("integer", 1)
-                                .put("string", "bar")
-                                .put("subConfig.integer", 2)
-                                .build(), 
-                            RuntimeLayer.class);
                 }
                 
                 @Provides
@@ -62,7 +60,7 @@ public class ProxyTest {
                 public MyConfig getMyConfig(ConfigProxyFactory factory) {
                     return factory.newProxy(MyConfig.class);
                 }
-            })
+            }
         );
         
         SettableConfig cfg = injector.getInstance(Key.get(SettableConfig.class, RuntimeLayer.class));
@@ -77,27 +75,24 @@ public class ProxyTest {
     }
     
     @Test
-    public void testConfigWithProvidedPrefix() {
-        Injector injector = Guice.createInjector(Modules
-            .override(new ArchaiusModule())
-            .with(new AbstractModule() {
-                @Override
-                protected void configure() {
-                    ConfigSeeders.bind(binder(), 
-                            MapConfig.builder()
-                                .put("prefix.integer", 1)
-                                .put("prefix.string", "bar")
-                                .build(), 
-                            RuntimeLayer.class);
-                }
-                
+    public void testConfigWithProvidedPrefix() throws ConfigException {
+        Injector injector = Guice.createInjector(
+            new ArchaiusModule()
+                .withApplicationOverrides(MapConfig.builder()
+                    .put("prefix.integer", 1)
+                    .put("prefix.string", "bar")
+                    .build()),
+            new AbstractModule() {
                 @Provides
                 @Singleton
                 public MyConfig getMyConfig(ConfigProxyFactory factory) {
                     return factory.newProxy(MyConfig.class, "prefix");
                 }
-            })
-            );
+
+                @Override
+                protected void configure() {
+                }
+            });
         
         MyConfig config = injector.getInstance(MyConfig.class);
         Assert.assertEquals("bar", config.getString());
