+++ archaius2-persisted2/src/test/java/com/netflix/archaius/persisted2/PlatformServiceTest.java
@@ -1,19 +1,16 @@
 package com.netflix.archaius.persisted2;
 
-import javax.inject.Singleton;
-
 import org.junit.Ignore;
 import org.junit.Test;
 
 import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
-import com.google.inject.Provides;
 import com.google.inject.Scopes;
-import com.google.inject.util.Modules;
 import com.netflix.archaius.api.Config;
-import com.netflix.archaius.guice.ArchaiusModule;
+import com.netflix.archaius.api.exceptions.ConfigException;
 import com.netflix.archaius.api.inject.RemoteLayer;
+import com.netflix.archaius.guice.ArchaiusModule;
 import com.netflix.archaius.visitor.PrintStreamVisitor;
 
 public class PlatformServiceTest {
@@ -21,7 +18,7 @@ public class PlatformServiceTest {
     // manually against internal Netflix systems
     @Test
     @Ignore
-    public void test() {
+    public void test() throws ConfigException {
         final Persisted2ClientConfig config = new DefaultPersisted2ClientConfig()
             .withServiceUrl("http://platformservice.us-east-1.dyntest.netflix.net:7001/platformservice/REST/v2/properties/jsonFilterprops")
             .withQueryScope("env",    "test")
@@ -34,19 +31,15 @@ public class PlatformServiceTest {
 //            .withSkipPropsWithExtraScopes(true)
             ;
 
-        Injector injector = Guice.createInjector(Modules.override(new ArchaiusModule()).with(new AbstractModule() {
-            @Override
-            protected void configure() {
-                bind(Persisted2ClientConfig.class).toInstance(config);
-                bind(Config.class).annotatedWith(RemoteLayer.class).toProvider(Persisted2ConfigProvider.class).in(Scopes.SINGLETON);
-            }
-            
-            @Provides
-            @Singleton
-            Config getConfig(@RemoteLayer Config config) {
-                return config;
-            }
-        }));
+        Injector injector = Guice.createInjector(
+                new ArchaiusModule(),
+                new AbstractModule() {
+                    @Override
+                    protected void configure() {
+                        bind(Persisted2ClientConfig.class).toInstance(config);
+                        bind(Config.class).annotatedWith(RemoteLayer.class).toProvider(Persisted2ConfigProvider.class).in(Scopes.SINGLETON);
+                    }
+                });
         
         Config c = injector.getInstance(Config.class);
         c.accept(new PrintStreamVisitor());
